+++ resources/js/app.ts
@@ -11,5 +11,6 @@ import './globals/chart';
 import './globals/lottie';
 import './globals/editor';
 import './globals/clipboard';
+import './globals/phone-input';
 import './components';